Enable stripe charge tests and refactor to reduce CodeClimate errors
